Move Element3D wrapper models into shared project. So they can be shared between WPF and UWP
