+++ Source/HelixToolkit.SharpDX.SharedModel/Element3D/BillboardTextModel3D.cs
@@ -2,11 +2,17 @@
 The MIT License (MIT)
 Copyright (c) 2018 Helix Toolkit contributors
 */
+#if NETFX_CORE
+using Windows.UI.Xaml;
 namespace HelixToolkit.UWP
+#else
+using System.Windows;
+namespace HelixToolkit.Wpf.SharpDX
+#endif
 {
     using Model;
     using Model.Scene;
-    using Windows.UI.Xaml;
+
     /// <summary>
     /// 
     /// </summary>
