+++ Source/HelixToolkit.SharpDX.SharedModel/Element3D/BoneSkinmeshGeometryModel3D.cs
@@ -1,17 +1,22 @@
 ﻿
 using SharpDX;
 using System.Collections.Generic;
-using System.Windows;
 
+#if NETFX_CORE
+using Windows.UI.Xaml;
+namespace HelixToolkit.UWP
+#else
+using System.Windows;
 namespace HelixToolkit.Wpf.SharpDX
+#endif
 {
     using Model;
     using Model.Scene;
 
     public class BoneSkinMeshGeometryModel3D : MeshGeometryModel3D
     {
-        public static DependencyProperty VertexBoneIdsProperty = DependencyProperty.Register("VertexBoneIds", typeof(IList<BoneIds>), typeof(BoneSkinMeshGeometryModel3D), 
-            new PropertyMetadata(null, (d,e)=>
+        public static DependencyProperty VertexBoneIdsProperty = DependencyProperty.Register("VertexBoneIds", typeof(IList<BoneIds>), typeof(BoneSkinMeshGeometryModel3D),
+            new PropertyMetadata(null, (d, e) =>
             {
                 ((d as Element3DCore).SceneNode as BoneSkinMeshNode).VertexBoneIds = e.NewValue as IList<BoneIds>;
             }));
@@ -29,7 +34,7 @@ namespace HelixToolkit.Wpf.SharpDX
         }
 
         public static DependencyProperty BoneMatricesProperty = DependencyProperty.Register("BoneMatrices", typeof(BoneMatricesStruct), typeof(BoneSkinMeshGeometryModel3D),
-            new PropertyMetadata(new BoneMatricesStruct() { Bones = new Matrix[BoneMatricesStruct.NumberOfBones] }, 
+            new PropertyMetadata(new BoneMatricesStruct() { Bones = new Matrix[BoneMatricesStruct.NumberOfBones] },
                 (d, e) =>
                 {
                     ((d as Element3DCore).SceneNode as BoneSkinMeshNode).BoneMatrices = (BoneMatricesStruct)e.NewValue;
@@ -54,7 +59,7 @@ namespace HelixToolkit.Wpf.SharpDX
 
         protected override void AssignDefaultValuesToSceneNode(SceneNode core)
         {
-            if(core is BoneSkinMeshNode n)
+            if (core is BoneSkinMeshNode n)
             {
                 n.BoneMatrices = BoneMatrices;
             }
