+++ Source/HelixToolkit.SharpDX.SharedModel/Element3D/EnvironmentMap3D.cs
@@ -1,7 +1,12 @@
-﻿using System.Windows;
-using System.IO;
+﻿using System.IO;
 
+#if NETFX_CORE
+using Windows.UI.Xaml;
+namespace HelixToolkit.UWP
+#else
+using System.Windows;
 namespace HelixToolkit.Wpf.SharpDX
+#endif
 {
     using Model;
     using Model.Scene;
@@ -10,7 +15,7 @@ namespace HelixToolkit.Wpf.SharpDX
     {
 
         public static readonly DependencyProperty TextureProperty = DependencyProperty.Register("Texture", typeof(Stream), typeof(EnvironmentMap3D),
-            new PropertyMetadata(null,(d,e)=> 
+            new PropertyMetadata(null, (d, e) =>
             {
                 ((d as Element3DCore).SceneNode as EnvironmentMapNode).Texture = (Stream)e.NewValue;
             }));
