+++ Source/HelixToolkit.SharpDX.SharedModel/Element3D/LineGeometryModel3D.cs
@@ -2,13 +2,20 @@
 The MIT License (MIT)
 Copyright (c) 2018 Helix Toolkit contributors
 */
+#if NETFX_CORE
+using Windows.UI.Xaml;
+using Media = Windows.UI;
 namespace HelixToolkit.UWP
+#else
+using System.Windows;
+using Media = System.Windows.Media;
+namespace HelixToolkit.Wpf.SharpDX
+#endif
 {
     using global::SharpDX;
     using Model;
     using Model.Scene;
-    using Windows.UI.Xaml;
-    using Media = Windows.UI;
+
     /// <summary>
     /// 
     /// </summary>
