+++ Source/HelixToolkit.SharpDX.SharedModel/Element3D/PointGeometryModel3D.cs
@@ -2,14 +2,24 @@
 The MIT License (MIT)
 Copyright (c) 2018 Helix Toolkit contributors
 */
+using global::SharpDX;
+#if NETFX_CORE
+using Windows.Foundation;
+using Windows.UI.Xaml;
+using Color = Windows.UI.Color;
+using Colors = Windows.UI.Colors;
+using Media = Windows.UI;
 namespace HelixToolkit.UWP
+#else
+using System.Windows;
+using Color = System.Windows.Media.Color;
+using Colors = System.Windows.Media.Colors;
+using Media = System.Windows.Media;
+namespace HelixToolkit.Wpf.SharpDX
+#endif
 {
-    using global::SharpDX;
     using Model;
-    using Model.Scene;
-    using Windows.Foundation;
-    using Windows.UI.Xaml;
-    using Media = Windows.UI;
+    using Model.Scene;      
 
     public class PointGeometryModel3D : GeometryModel3D
     {
