+++ Source/HelixToolkit.Wpf.SharpDX/HelixToolkit.Wpf.SharpDX.csproj
@@ -200,24 +200,12 @@
     <Compile Include="Model\Elements2D\Abstract\Element2D.cs" />
     <Compile Include="Model\Elements2D\Canvas2D.cs" />
     <Compile Include="Model\Elements3D\CameraModel3D.cs" />
-    <Compile Include="Model\Elements3D\CrossSectionMeshGeometryModel3D.cs" />
     <Compile Include="Model\Elements3D\AbstractElements3D\ScreenSpaceMeshGeometry3D.cs" />
-    <Compile Include="Model\Elements3D\BillboardTextModel3D.cs" />
-    <Compile Include="Model\Elements3D\BoneSkinMeshGeometryModel3D.cs" />
-    <Compile Include="Model\Elements3D\CoordinateSystemModel3D.cs" />
     <Compile Include="Model\Elements3D\Element3DPresenter.cs" />
-    <Compile Include="Model\Elements3D\InstancingBillboardModel3D.cs" />
     <Compile Include="Model\Elements3D\InstancingMeshGeometryModel3D.cs" />
     <Compile Include="Model\Elements3D\OctreeLineGeometryModel3D.cs" />
     <Compile Include="Model\Elements3D\OutLineMeshGeometryModel3D.cs" />
     <Compile Include="Model\Elements3D\ParticleStormModel3D.cs" />
-    <Compile Include="Model\Elements3D\PatchGeometryModel3D.cs" />
-    <Compile Include="Model\Elements3D\PointGeometryModel3D.cs" />
-    <Compile Include="Model\Elements3D\EnvironmentMap3D.cs" />
-    <Compile Include="Model\Elements3D\PostEffects\PostEffectBloom.cs" />
-    <Compile Include="Model\Elements3D\PostEffects\PostEffectMeshBorderHighlight.cs" />
-    <Compile Include="Model\Elements3D\PostEffects\PostEffectMeshOutlineBlur.cs" />
-    <Compile Include="Model\Elements3D\PostEffects\PostEffectMeshXRay.cs" />
     <Compile Include="Model\Elements3D\ScreenDuplicationModel.cs" />
     <Compile Include="Model\Elements3D\ViewBoxModel3D.cs" />
     <Compile Include="Model\Materials\Material.cs" />
@@ -249,7 +237,6 @@
     <Compile Include="Model\Elements3D\GroupModel3D.cs" />
     <Compile Include="Model\Elements3D\ItemsModel3D.cs" />
     <Compile Include="Model\Elements3D\Light3DCollection.cs" />
-    <Compile Include="Model\Elements3D\MeshGeometryModel3D.cs" />
     <Compile Include="Model\Elements3D\Element3DCollection.cs" />
     <Compile Include="Model\Elements3D\AbstractElements3D\MaterialGeometryModel3D.cs" />
     <Compile Include="Model\Elements3D\UICompositeManipulator3D.cs" />
@@ -263,7 +250,6 @@
     <Compile Include="Model\Elements3D\AbstractElements3D\GeometryModel3D.cs" />
     <Compile Include="Model\Elements3D\AbstractElements3D\GroupElement3D.cs" />
     <Compile Include="Model\Lights3D\Light3D.cs" />
-    <Compile Include="Model\Elements3D\LineGeometryModel3D.cs" />
     <Compile Include="Model\Lights3D\PointLight3D.cs" />
     <Compile Include="Model\Lights3D\ShadowMap3D.cs" />
     <Compile Include="Model\Lights3D\SpotLight3D.cs" />
@@ -386,5 +372,6 @@
   <ItemGroup />
   <Import Project="..\HelixToolkit.SharpDX.Shared\HelixToolkit.SharpDX.Shared.projitems" Label="Shared" />
   <Import Project="..\HelixToolkit.Shared\HelixToolkit.Shared.projitems" Label="Shared" />
+  <Import Project="..\HelixToolkit.SharpDX.SharedModel\HelixToolkit.SharpDX.SharedModel.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
