--- Source/HelixToolkit.Wpf.SharpDX/Model/Elements3D/BillboardTextModel3D.cs
@@ -1,92 +1,0 @@
-﻿using SharpDX;
-using SharpDX.Direct3D11;
-using System;
-using System.Collections.Generic;
-using System.Windows;
-
-namespace HelixToolkit.Wpf.SharpDX
-{
-    using Model;
-    using Model.Scene;
-    /// <summary>
-    /// 
-    /// </summary>
-    /// <seealso cref="HelixToolkit.Wpf.SharpDX.GeometryModel3D" />
-    public class BillboardTextModel3D : GeometryModel3D
-    {
-        #region Dependency Properties
-        /// <summary>
-        /// Fixed sized billboard. Default = true. 
-        /// <para>When FixedSize = true, the billboard render size will be scale to normalized device coordinates(screen) size</para>
-        /// <para>When FixedSize = false, the billboard render size will be actual size in 3D world space</para>
-        /// </summary>
-        public static readonly DependencyProperty FixedSizeProperty = DependencyProperty.Register("FixedSize", typeof(bool), typeof(BillboardTextModel3D),
-            new PropertyMetadata(true,
-                (d, e) =>
-                {
-                    ((d as Element3DCore).SceneNode as BillboardNode).FixedSize = (bool)e.NewValue;
-                }));
-
-        /// <summary>
-        /// Fixed sized billboard. Default = true. 
-        /// <para>When FixedSize = true, the billboard render size will be scale to normalized device coordinates(screen) size</para>
-        /// <para>When FixedSize = false, the billboard render size will be actual size in 3D world space</para>
-        /// </summary>
-        public bool FixedSize
-        {
-            set
-            {
-                SetValue(FixedSizeProperty, value);
-            }
-            get
-            {
-                return (bool)GetValue(FixedSizeProperty);
-            }
-        }
-
-        /// <summary>
-        /// Specifiy if billboard texture is transparent. 
-        /// During rendering, transparent objects are rendered after opaque objects. Transparent objects' order in scene graph are preserved.
-        /// </summary>
-        public static readonly DependencyProperty IsTransparentProperty =
-            DependencyProperty.Register("IsTransparent", typeof(bool), typeof(BillboardTextModel3D), new PropertyMetadata(false, (d, e) =>
-            {
-                ((d as Element3DCore).SceneNode as BillboardNode).IsTransparent = (bool)e.NewValue;
-            }));
-
-        /// <summary>
-        /// Specifiy if  billboard texture is transparent. 
-        /// During rendering, transparent objects are rendered after opaque objects. Transparent objects' order in scene graph are preserved.
-        /// </summary>
-        public bool IsTransparent
-        {
-            get { return (bool)GetValue(IsTransparentProperty); }
-            set { SetValue(IsTransparentProperty, value); }
-        }
-        #endregion
-
-        #region Overridable Methods        
-
-        /// <summary>
-        /// Called when [create scene node].
-        /// </summary>
-        /// <returns></returns>
-        protected override SceneNode OnCreateSceneNode()
-        {
-            return new BillboardNode();
-        }
-        /// <summary>
-        /// Assigns the default values to core.
-        /// </summary>
-        /// <param name="core">The core.</param>
-        protected override void AssignDefaultValuesToSceneNode(SceneNode core)
-        {
-            if (core is BillboardNode n)
-            {
-                n.FixedSize = FixedSize;
-            }
-            base.AssignDefaultValuesToSceneNode(core);       
-        }
-        #endregion
-    }
-}
