--- Source/HelixToolkit.Wpf.SharpDX/Model/Elements3D/PointGeometryModel3D.cs
@@ -1,108 +1,0 @@
-﻿namespace HelixToolkit.Wpf.SharpDX
-{
-    using global::SharpDX;
-    using Model;
-    using Model.Scene;
-    using System.Windows;
-    using Media = System.Windows.Media;
-
-    public class PointGeometryModel3D : GeometryModel3D
-    {
-        #region Dependency Properties
-        public static readonly DependencyProperty ColorProperty =
-            DependencyProperty.Register("Color", typeof(Media.Color), typeof(PointGeometryModel3D),
-                new PropertyMetadata(Media.Colors.Black, (d, e) =>
-                {
-                    ((d as Element3DCore).SceneNode as PointNode).Color = ((Media.Color)e.NewValue).ToColor4();
-                }));
-
-        public static readonly DependencyProperty SizeProperty =
-            DependencyProperty.Register("Size", typeof(Size), typeof(PointGeometryModel3D), new PropertyMetadata(new Size(1.0, 1.0),
-                (d,e)=> 
-                {
-                    var size = (Size)e.NewValue;
-                    ((d as Element3DCore).SceneNode as PointNode).Size = new Size2F((float)size.Width, (float)size.Height);
-                }));
-
-        public static readonly DependencyProperty FigureProperty =
-            DependencyProperty.Register("Figure", typeof(PointFigure), typeof(PointGeometryModel3D), new PropertyMetadata(PointFigure.Rect,
-                (d, e)=> 
-                {
-                    ((d as Element3DCore).SceneNode as PointNode).Figure = (PointFigure)e.NewValue;
-                }));
-
-        public static readonly DependencyProperty FigureRatioProperty =
-            DependencyProperty.Register("FigureRatio", typeof(double), typeof(PointGeometryModel3D), new PropertyMetadata(0.25,
-                (d, e)=> 
-                {
-                    ((d as Element3DCore).SceneNode as PointNode).FigureRatio = (float)(double)e.NewValue;
-                }));
-
-        public static readonly DependencyProperty HitTestThicknessProperty =
-            DependencyProperty.Register("HitTestThickness", typeof(double), typeof(PointGeometryModel3D), new PropertyMetadata(4.0, (d, e)=> 
-                {
-                    ((d as Element3DCore).SceneNode as PointNode).HitTestThickness = (float)(double)e.NewValue;
-                }));
-
-        public Media.Color Color
-        {
-            get { return (Media.Color)this.GetValue(ColorProperty); }
-            set { this.SetValue(ColorProperty, value); }
-        }
-
-        public Size Size
-        {
-            get { return (Size)this.GetValue(SizeProperty); }
-            set { this.SetValue(SizeProperty, value); }
-        }
-
-        public PointFigure Figure
-        {
-            get { return (PointFigure)this.GetValue(FigureProperty); }
-            set { this.SetValue(FigureProperty, value); }
-        }
-
-        public double FigureRatio
-        {
-            get { return (double)this.GetValue(FigureRatioProperty); }
-            set { this.SetValue(FigureRatioProperty, value); }
-        }
-
-        /// <summary>
-        /// Used only for point/line hit test
-        /// </summary>
-        public double HitTestThickness
-        {
-            get { return (double)this.GetValue(HitTestThicknessProperty); }
-            set { this.SetValue(HitTestThicknessProperty, value); }
-        }
-        #endregion
-
-
-        /// <summary>
-        /// Called when [create render core].
-        /// </summary>
-        /// <returns></returns>
-        protected override SceneNode OnCreateSceneNode()
-        {
-            return new PointNode();
-        }
-        /// <summary>
-        /// Assigns the default values to core.
-        /// </summary>
-        /// <param name="core">The core.</param>
-        protected override void AssignDefaultValuesToSceneNode(SceneNode core)
-        {
-            if (core is PointNode n)
-            {
-                n.Size = new Size2F((float)Size.Width, (float)Size.Height);
-                n.Figure = Figure;
-                n.FigureRatio = (float)FigureRatio;
-                n.Color = Color.ToColor4();
-            }
-
-            base.AssignDefaultValuesToSceneNode(core);
-        }
-    }
-
-}
